app/tasks/api/serializers.py
from rest_framework import serializers

from app.tasks.models import Task


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ["id", "title", "description", "completed"]
